refactor: clean up calltaker code, add expandable-itinerary
