+++ lib/actions/call-taker.js
@@ -3,7 +3,6 @@ import qs from 'qs'
 import { createAction } from 'redux-actions'
 
 import {getRoutingParams} from '../actions/api'
-// import {getCallForId} from '../util/call-taker'
 import {URL_ROOT} from '../util/constants'
 import {getTimestamp} from '../util/state'
 
@@ -12,16 +11,20 @@ if (typeof (fetch) === 'undefined') require('isomorphic-fetch')
 /// PRIVATE ACTIONS
 
 const endingCall = createAction('END_CALL')
-const requestingCalls = createAction('REQUESTING_CALLS')
 const receivedCalls = createAction('RECEIVED_CALLS')
-const requestingQueries = createAction('REQUESTING_QUERIES')
 const receivedQueries = createAction('RECEIVED_QUERIES')
+const requestingCalls = createAction('REQUESTING_CALLS')
+const requestingQueries = createAction('REQUESTING_QUERIES')
+const storeSession = createAction('STORE_SESSION')
 
 /// PUBLIC ACTIONS
 
 export const beginCall = createAction('BEGIN_CALL')
 export const toggleCallHistory = createAction('TOGGLE_CALL_HISTORY')
 
+/**
+ * End the active call and store the queries made during the call.
+ */
 export function endCall () {
   return function (dispatch, getState) {
     const {callTaker, otp} = getState()
@@ -51,12 +54,16 @@ export function endCall () {
   }
 }
 
+/**
+ * Initialize the Call Taker and Field Trip modules by checking the session
+ * query param against sessions in the datastore backend or initializing a new
+ * session via Trinet.
+ */
 export function initializeModules () {
   return function (dispatch, getState) {
     const {datastoreUrl, trinetReDirect} = getState().otp.config
     // Initialize session if datastore enabled.
     if (datastoreUrl && trinetReDirect) {
-      // TODO: Use session storage?
       // TODO: Generalize for non-TriNet instances.
       const sessionId = getUrlParams().sessionId
       if (sessionId) {
@@ -89,9 +96,11 @@ function newSession (datastoreUrl, verifyLoginUrl, redirect) {
     })
 }
 
-const storeSession = createAction('STORE_SESSION')
-
-function checkSession (datastoreUrl, sessionId, checkSessionSuccessCallback) {
+/**
+ * Check that a particular session ID is valid and store resulting session
+ * data.
+ */
+function checkSession (datastoreUrl, sessionId) {
   return function (dispatch, getState) {
     fetch(datastoreUrl + `/auth/checkSession?sessionId=${sessionId}`)
       .then(res => res.json())
@@ -103,6 +112,9 @@ function checkSession (datastoreUrl, sessionId, checkSessionSuccessCallback) {
   }
 }
 
+/**
+ * Fetch latest calls for a particular session.
+ */
 export function fetchCalls () {
   return function (dispatch, getState) {
     dispatch(requestingCalls())
@@ -127,6 +139,9 @@ function placeToLatLonStr (place) {
   return `${place.lat.toFixed(6)},${place.lon.toFixed(6)}`
 }
 
+/**
+ * @return {boolean} - whether a calltaker session is invalid
+ */
 function sessionIsInvalid (session) {
   if (!session || !session.sessionId) {
     console.error('No valid OTP datastore session found.')
@@ -135,6 +150,10 @@ function sessionIsInvalid (session) {
   return false
 }
 
+/**
+ * Store the trip queries made over the course of a call (to be called when the
+ * call terminates).
+ */
 export function saveQueriesForCall (call) {
   return function (dispatch, getState) {
     const {callTaker, otp} = getState()
@@ -168,6 +187,9 @@ export function saveQueriesForCall (call) {
   }
 }
 
+/**
+ * Fetch the trip queries that were made during a particular call.
+ */
 export function fetchQueries (callId) {
   return function (dispatch, getState) {
     dispatch(requestingQueries())
