+++ lib/actions/form.js
@@ -85,7 +85,6 @@ export function formChanged (oldQuery, newQuery) {
 
     // If departArrive is set to 'NOW', update the query time to current
     if (otpState.currentQuery && otpState.currentQuery.departArrive === 'NOW') {
-      console.log(newQuery)
       dispatch(settingQueryParam({ time: moment().format(coreUtils.time.OTP_API_TIME_FORMAT) }))
     }
