+++ lib/actions/ui.js
@@ -7,6 +7,7 @@ import { findRoute } from './api'
 import { setMapCenter, setMapZoom, setRouterId } from './config'
 import { clearActiveSearch, parseUrlQueryString, setActiveSearch } from './form'
 import { clearLocation } from './map'
+import { setActiveItinerary } from './narrative'
 import { getUiUrlParams } from '../util/state'
 
 /**
@@ -122,8 +123,7 @@ export function handleBackButtonPress (e) {
         dispatch(setActiveSearch(previousSearchId))
       } else if (uiUrlParams.ui_activeItinerary !== previousItinIndex) {
         // Active itinerary index has changed.
-        console.log('hidden setActiveItinerary!')
-        // dispatch(setActiveItinerary({ index: previousItinIndex }))
+        dispatch(setActiveItinerary({ index: previousItinIndex }))
       }
     } else {
       // The back button was pressed, but there was no corresponding search
