+++ lib/components/admin/call-record.js
@@ -6,6 +6,9 @@ import * as formActions from '../../actions/form'
 import CallTimeCounter from './call-time-counter'
 import Icon from '../narrative/icon'
 
+/**
+ * Displays information for a particular call record in the Call Taker window.
+ */
 export default class CallRecord extends Component {
   state = {
     expanded: false
