+++ lib/components/admin/call-taker-windows.js
@@ -7,16 +7,11 @@ import * as callTakerActions from '../../actions/call-taker'
 import CallRecord from './call-record'
 import Icon from '../narrative/icon'
 
+/**
+ * Collects the various draggable windows used in the Call Taker module to
+ * display, for example, the call record list and (TODO) the list of field trips.
+ */
 class CallTakerWindows extends Component {
-  // FIXME props
-  static propTypes = {
-    routingType: PropTypes.string,
-    text: PropTypes.string,
-    onClick: PropTypes.func,
-    planTrip: PropTypes.func,
-    profileTrip: PropTypes.func
-  }
-
   _onCloseCallHistory = () => this.props.toggleCallHistory()
 
   render () {
