+++ lib/components/app/call-taker-panel.js
@@ -19,6 +19,7 @@ import IntermediatePlace from '../form/intermediate-place-field'
 import LocationField from '../form/connected-location-field'
 import SwitchButton from '../form/switch-button'
 import UserSettings from '../form/user-settings'
+import ExpandableItinerary from '../narrative/expandable-itinerary'
 import Icon from '../narrative/icon'
 import NarrativeItineraries from '../narrative/narrative-itineraries'
 import { getActiveSearch, getShowUserSettings } from '../../util/state'
@@ -90,6 +91,11 @@ const modeOptions = [
 const metersToMiles = meters => Math.round(meters * 0.000621371 * 100) / 100
 const milesToMeters = miles => miles / 0.000621371
 
+/**
+ * This is the main panel/sidebar for the Call Taker/Field Trip module. It
+ * currently also serves as the main panel for the FDOT RMCE trip comparison view
+ * which depends on the BATCH trip planning mode.
+ */
 class CallTakerPanel extends Component {
   constructor (props) {
     super(props)
@@ -211,7 +217,6 @@ class CallTakerPanel extends Component {
       activeSearch,
       datastoreUrl,
       currentQuery,
-      itineraryClass,
       itineraryFooter,
       LegIcon,
       mainPanelContent,
@@ -386,7 +391,7 @@ class CallTakerPanel extends Component {
           }
           <div className='desktop-narrative-container'>
             <NarrativeItineraries
-              itineraryClass={itineraryClass}
+              itineraryClass={ExpandableItinerary}
               itineraryFooter={itineraryFooter}
               LegIcon={LegIcon}
             />
