+++ lib/components/narrative/default/default-itinerary.js
@@ -2,178 +2,50 @@ import coreUtils from '@opentripplanner/core-utils'
 import React from 'react'
 
 import NarrativeItinerary from '../narrative-itinerary'
-import ItineraryBody from '../line-itin/connected-itinerary-body'
 import ItinerarySummary from './itinerary-summary'
-import { getTotalFareAsString } from '../../../util/state'
+import ItineraryDetails from './itinerary-details'
+import TripDetails from '../connected-trip-details'
+import TripTools from '../trip-tools'
 
-const { isBicycle, isTransit } = coreUtils.itinerary
 const { formatDuration, formatTime } = coreUtils.time
 
-// FIXME move to core utils
-function getItineraryDescription (itinerary) {
-  let primaryTransitDuration = 0
-  let mainMode = 'Walk'
-  let transitMode
-  itinerary.legs.forEach((leg, i) => {
-    const {duration, mode, rentedBike} = leg
-    if (isTransit(mode) && duration > primaryTransitDuration) {
-      primaryTransitDuration = duration
-      transitMode = mode.toLowerCase()
-    }
-    if (isBicycle(mode)) mainMode = 'Bike'
-    if (rentedBike) mainMode = 'Bikeshare'
-    if (mode === 'CAR') mainMode = 'Drive'
-  })
-  let description = mainMode
-  if (transitMode) description += ` to ${transitMode}`
-  return description
-}
-
-const ITINERARY_ATTRIBUTES = [
-  {
-    alias: 'best',
-    id: 'duration',
-    order: 0,
-    render: (itinerary, options) => formatDuration(itinerary.duration)
-  },
-  {
-    alias: 'departureTime',
-    id: 'arrivalTime',
-    order: 1,
-    render: (itinerary, options) => {
-      if (options.isSelected) {
-        if (options.selection === 'ARRIVALTIME') return formatTime(itinerary.endTime, options)
-        else return formatTime(itinerary.startTime, options)
-      }
-      return (
-        <span>
-          {formatTime(itinerary.startTime, options)}
-          —
-          {formatTime(itinerary.endTime, options)}
-        </span>
-      )
-    }
-  },
-  {
-    id: 'cost',
-    order: 2,
-    render: (itinerary, options) => getTotalFareAsString(itinerary)
-  },
-  {
-    id: 'walkTime',
-    order: 3,
-    render: (itinerary, options) => {
-      const leg = itinerary.legs[0]
-      const {LegIcon} = options
-      return (
-        // FIXME: For CAR mode, walk time considers driving time.
-        <span>
-          {formatDuration(itinerary.walkTime)}{' '}
-          <div style={{
-            width: '20px',
-            height: '20px',
-            position: 'absolute',
-            display: 'inline',
-            paddingLeft: '4px',
-            paddingBottom: '4px'
-          }}>
-            <LegIcon leg={leg} />
-          </div>
-        </span>
-      )
-    }
-  }
-]
-
 export default class DefaultItinerary extends NarrativeItinerary {
-  _onMouseOver = () => {
-    const {index} = this.props
-    // FIXME: Need to add focus call with ref?
-    this.props.setVisibleItinerary({index})
-  }
-
-  _onMouseOut = () => {
-    this.props.setVisibleItinerary({index: null})
-  }
-
-  _isSortingOnAttribute = (attribute) => {
-    const {sort} = this.props
-    const type = sort.type.toLowerCase()
-    return attribute.id.toLowerCase() === type ||
-      (attribute.alias && attribute.alias.toLowerCase() === type)
-  }
-
   render () {
     const {
       active,
+      activeLeg,
+      activeStep,
       expanded,
+      index,
       itinerary,
       LegIcon,
-      timeFormat
+      setActiveLeg,
+      setActiveStep
     } = this.props
-    const timeOptions = {
-      format: timeFormat,
-      offset: coreUtils.itinerary.getTimeZoneOffset(itinerary)
-    }
     return (
-      <div
-        className={`option default-itin${active ? ' active' : ''}${expanded ? ' expanded' : ''}`}
-        role='presentation'
-        // FIXME: Move style to css
-        style={{
-          borderLeft: active && !expanded ? '4px teal solid' : undefined,
-          backgroundColor: expanded ? 'white' : undefined
-        }}
-        // TODO: Somehow set first itinerary as visible initially.
-        // autoFocus={index === 0}
-        onBlur={this._onMouseOut}
-        onFocus={this._onMouseOver}
-        onMouseOut={this._onMouseOut}
-        onMouseOver={this._onMouseOver}
-      >
+      <div className={`option default-itin${active ? ' active' : ''}`}>
         <button
           className='header'
-          // _onHeaderClick comes from super component (NarrativeItinerary).
           onClick={this._onHeaderClick}
         >
-          <div
-            className='title'>
-            {getItineraryDescription(itinerary)}
-          </div>
-          <ul className='list-unstyled itinerary-attributes'>
-            {ITINERARY_ATTRIBUTES
-              .sort((a, b) => {
-                const aSelected = this._isSortingOnAttribute(a)
-                const bSelected = this._isSortingOnAttribute(b)
-                if (aSelected) return -1
-                if (bSelected) return 1
-                else return a.order - b.order
-              })
-              .map(attribute => {
-                const isSelected = this._isSortingOnAttribute(attribute)
-                const options = attribute.id === 'arrivalTime' ? timeOptions : {}
-                if (isSelected) {
-                  options.isSelected = true
-                  options.selection = this.props.sort.type
-                }
-                options.LegIcon = LegIcon
-                return (
-                  <li key={attribute.id} className={`${attribute.id}${isSelected ? ' main' : ''}`}>
-                    {attribute.render(itinerary, options)}
-                  </li>
-                )
-              })
-            }
-          </ul>
+          <span className='title'>Itinerary {index + 1}</span>{' '}
+          <span className='duration pull-right'>{formatDuration(itinerary.duration)}</span>{' '}
+          <span className='arrivalTime'>{formatTime(itinerary.startTime)}—{formatTime(itinerary.endTime)}</span>
           <ItinerarySummary itinerary={itinerary} LegIcon={LegIcon} />
-          {(active && !expanded) &&
-            <small style={{clear: 'both', textAlign: 'center', display: 'block', color: 'grey'}}>
-              click to view details
-            </small>
-          }
         </button>
-        {(active && expanded) &&
-          <ItineraryBody itinerary={itinerary} LegIcon={LegIcon} />
+        {(active || expanded) &&
+          <div className='body'>
+            <ItineraryDetails
+              itinerary={itinerary}
+              activeLeg={activeLeg}
+              activeStep={activeStep}
+              setActiveLeg={setActiveLeg}
+              setActiveStep={setActiveStep}
+              LegIcon={LegIcon}
+            />
+            <TripDetails itinerary={itinerary} />
+            <TripTools itinerary={itinerary} />
+          </div>
         }
       </div>
     )
