+++ lib/components/narrative/narrative-itineraries.js
@@ -174,7 +174,6 @@ class NarrativeItineraries extends Component {
             const active = index === activeItinerary
             // Hide non-active itineraries.
             if (!active && itineraryIsExpanded) return null
-            if (active) console.log(itinerary)
             return React.createElement(itineraryClass, {
               itinerary,
               index,
@@ -201,7 +200,6 @@ const mapStateToProps = (state, ownProps) => {
   const {persistence} = state.otp.config
   const {modes} = state.otp.config
   const {filter, sort} = state.otp.filter
-  // const { activeItinerary, activeLeg, activeStep } = activeSearch ? activeSearch.activeItinerary : {}
   const pending = activeSearch ? Boolean(activeSearch.pending) : false
   const itineraries = getActiveItineraries(state.otp)
   const realtimeEffects = getRealtimeEffects(state.otp)
