+++ lib/components/narrative/narrative-routing-results.js
@@ -44,11 +44,12 @@ class NarrativeRoutingResults extends Component {
 
 const mapStateToProps = (state, ownProps) => {
   const activeSearch = getActiveSearch(state.otp)
+  const pending = activeSearch ? Boolean(activeSearch.pending) : false
   return {
     mainPanelContent: state.otp.ui.mainPanelContent,
     error: activeSearch && activeSearch.response && activeSearch.response.error,
     itineraries: getActiveItineraries(state.otp),
-    pending: activeSearch && activeSearch.pending,
+    pending,
     routingType: activeSearch && activeSearch.query.routingType
   }
 }
