+++ lib/components/narrative/tabbed-itineraries.js
@@ -51,7 +51,6 @@ class TabbedItineraries extends Component {
         realtimeEffects.routesDiffer ||
         !useRealtime
       )
-
     return (
       <div className='options itinerary tabbed-itineraries'>
         <div className='tab-row'>
@@ -116,16 +115,19 @@ class TabbedItineraries extends Component {
           useRealtime={useRealtime} />
         */}
 
-        {/* Show the active itinerary */}
-        {(activeItinerary !== null) && React.createElement(itineraryClass, {
-          itinerary: itineraries[activeItinerary],
-          index: activeItinerary,
-          key: activeItinerary,
-          active: true,
-          routingType: 'ITINERARY',
-          showRealtimeAnnotation,
-          ...this.props
-        })}
+        {/* Show active itin if itineraries exist and active itin is defined. */}
+        {(itineraries.length > 0 && activeItinerary >= 0)
+          ? React.createElement(itineraryClass, {
+            itinerary: itineraries[activeItinerary],
+            index: activeItinerary,
+            key: activeItinerary,
+            active: true,
+            routingType: 'ITINERARY',
+            showRealtimeAnnotation,
+            ...this.props
+          })
+          : null
+        }
 
       </div>
     )
